Important Type Conversion Fix

+++ script.py
@@ -116,7 +116,7 @@ for headline in clean_headlines:
 for i in range(0, len(clean_headlines)):
   for j in range(0, len(models)):
     # Flag any labels with a confidence score < 60%
-    if (int(response_data[i][j][1]) < 0.6):
+    if (float(response_data[i][j][1]) < 0.6):
       response_data[i][j][1] += " - BELOW THRESHOLD"
 
 for i in range(0, len(clean_headlines)):
